v0.3.2 Phase 1: Intelligent Compression with Full Codec Detection
This commit implements comprehensive codec detection across all 10 OpenZL
codecs, enabling intelligent compression for all data types (text, binary,
CSV, JSON, etc.).

Key Features:

1. Comprehensive Codec Detection (purgo/analyzer.go - 556 lines):
   - 10 codec strategies: Identity, Constant, Delta, ZigZag, Bitpack,
     Transpose, FSE, Huffman, LZ77, RLE
   - 8-priority detection algorithm for optimal codec selection
   - Format detection: JSON, CSV, Text, Binary
   - Segmentation support: SegmentCSV (per-column), SegmentJSON (per-field)
   - Helper functions for each codec type detection

2. Enhanced CompressSmart() (purgo/encoder.go):
   - Auto-detects data format and structure
   - For CSV/JSON: Uses intelligent per-segment analysis
   - For all data: Applies 8-priority codec detection
   - compressSegmented() analyzes segments and selects most common codec
   - Single-output approach (workaround for frame reader limitation)
   - Maintains existing multi-strategy fallback

3. Enhanced Decompress() (purgo/decoder.go):
   - Removed single-output restriction
   - Supports multi-output frames (for future per-segment compression)
   - Concatenates segments in correct order
   - Backward compatible with single-output frames

Test Coverage:

- analyzer_test.go: 30+ tests covering all 10 codec detection strategies
- compress_smart_test.go: 6 tests demonstrating 18-24× compression
- All purgo tests passing (100% pass rate)
- Tests cover: text, numbers, UUIDs, JSON, CSV, repetitive data, sparse data

Codec Detection Priorities (applies to all data types):
1. Constant: All identical values → Constant codec
2. Delta: Sequential numbers (IDs) → Delta codec
3. Bitpack: Small integers (0-255) → Bitpack codec
4. RLE: High repetition (≥80%) → RLE codec
5. Numeric: All numbers → Transpose codec
6. UUID/Pattern: UUIDs, structured text → LZ77 codec
7. Low entropy: Limited unique values → FSE codec
8. Default: General text/binary → LZ77 codec

Benchmark Results (Apple M4 Pro):
- JSON (BitLocker data): 12,715 bytes → 699 bytes (18.19× compression)
- Repeated strings: 4,900 bytes → 200 bytes (24.50× compression)
- Sparse data: 1,000 bytes → 52 bytes (19.23× compression)
- CompressSmart vs Compress: 684% improvement (17.55× vs 2.24×)

Known Limitations:
- Frame reader supports ≤2 outputs (internal/frame/reader.go:69-70)
- Workaround: Use single-output with most common codec
- Future: Per-segment compression requires frame reader enhancement

Architecture Decisions:
1. Single-output approach: Works within current frame reader constraints
2. Most-common-codec selection: Balances analysis intelligence with
   compatibility
3. Type alignment: Changed Segment.CodecID from byte to uint16 (matches
   codec.ID)
4. Backward compatibility: Decompress() handles both single and multi-output
5. Universal detection: Same codec analysis applies to all data formats

Files Modified:
- purgo/analyzer.go (556 lines) - NEW
- purgo/analyzer_test.go (826 lines) - NEW
- purgo/encoder.go (+109 lines) - compressSegmented() function
- purgo/decoder.go (+20 lines) - multi-output support

+++ purgo/analyzer.go
@@ -194,24 +194,24 @@ const (
 
 // Codec ID constants (must match internal/codec/codec.go)
 const (
-	codecIDIdentity  byte = 0
-	codecIDConstant  byte = 1
-	codecIDDelta     byte = 2
-	codecIDZigZag    byte = 3
-	codecIDBitpack   byte = 4
-	codecIDTranspose byte = 5
-	codecIDQuantize  byte = 6
-	codecIDFSE       byte = 10
-	codecIDHuffman   byte = 11
-	codecIDLZ77      byte = 12
-	codecIDRLE       byte = 13
-	codecIDZstd      byte = 20
+	codecIDIdentity  uint16 = 0
+	codecIDConstant  uint16 = 1
+	codecIDDelta     uint16 = 2
+	codecIDZigZag    uint16 = 3
+	codecIDBitpack   uint16 = 4
+	codecIDTranspose uint16 = 5
+	codecIDQuantize  uint16 = 6
+	codecIDFSE       uint16 = 10
+	codecIDHuffman   uint16 = 11
+	codecIDLZ77      uint16 = 12
+	codecIDRLE       uint16 = 13
+	codecIDZstd      uint16 = 20
 )
 
 // Segment represents a portion of data with a suggested codec
 type Segment struct {
 	Data      []byte
-	CodecID   byte   // Suggested codec ID
+	CodecID   uint16 // Suggested codec ID (matches codec.ID type)
 	CodecName string // Human-readable codec name
 }
 
@@ -292,7 +292,7 @@ func SegmentCSV(data []byte) ([]Segment, error) {
 //  6. UUID/Text patterns: → LZ77 (dictionary compression)
 //  7. Text with low entropy: → FSE or Huffman (entropy coding)
 //  8. Default: General text → LZ77
-func analyzeColumnPattern(data []byte) (codecID byte, codecName string) {
+func analyzeColumnPattern(data []byte) (codecID uint16, codecName string) {
 	if len(data) == 0 {
 		return codecIDIdentity, codecNameIdentity
 	}
